+++ frontend/src/App.vue
@@ -2,22 +2,33 @@
 import Notification from "@/components/common/Notification.vue";
 import api from "@/services/api/index";
 import socket from "@/services/socket";
+import storeAuth from "@/stores/auth";
 import storeConfig from "@/stores/config";
 import storeGalleryFilter from "@/stores/galleryFilter";
 import storeHeartbeat from "@/stores/heartbeat";
-import storeNotifications from "@/stores/notifications";
+import RommIso from "@/components/common/RommIso.vue";
 import storeRoms, { type SimpleRom } from "@/stores/roms";
 import storeScanning from "@/stores/scanning";
+import storeNavigation from "@/stores/navigation";
 import type { Events } from "@/types/emitter";
-import { normalizeString } from "@/utils";
+import { defaultAvatarPath, normalizeString } from "@/utils";
 import type { Emitter } from "mitt";
 import { storeToRefs } from "pinia";
-import { inject, onBeforeUnmount, onMounted } from "vue";
+import { inject, onBeforeUnmount, onMounted, ref } from "vue";
+import { useRouter } from "vue-router";
+import { useDisplay } from "vuetify";
 
 // Props
+const { mdAndDown } = useDisplay();
+const router = useRouter();
+const auth = storeAuth();
+const platformsDrawer = ref(false);
+const settingsDrawer = ref(false);
+const actionsDrawer = ref(false);
+const navigationStore = storeNavigation();
+const { activeSettingsDrawer, activePlatformsDrawer } = storeToRefs(navigationStore);
 const scanningStore = storeScanning();
-const notificationStore = storeNotifications();
-const { notifications } = storeToRefs(notificationStore);
+const { scanning } = storeToRefs(scanningStore);
 const { scanningPlatforms } = storeToRefs(scanningStore);
 const romsStore = storeRoms();
 const galleryFilter = storeGalleryFilter();
@@ -89,6 +100,33 @@ socket.on("scan:done_ko", (msg) => {
   socket.disconnect();
 });
 
+// Functions
+function goHome() {
+  navigationStore.switchActivePlatformsDrawer(false)
+  navigationStore.switchActiveSettingsDrawer(false)
+  router.push({ name: "dashboard" });
+}
+function goScan() {
+  navigationStore.switchActivePlatformsDrawer(false)
+  navigationStore.switchActiveSettingsDrawer(false)
+  router.push({ name: "dashboard" });
+}
+function togglePlatformsDrawer() {
+  platformsDrawer.value = !platformsDrawer.value;
+  settingsDrawer.value = false;
+  actionsDrawer.value = false;
+}
+function toggleSettingsDrawer() {
+  platformsDrawer.value = false;
+  settingsDrawer.value = !settingsDrawer.value;
+  actionsDrawer.value = false;
+}
+function toggleActionsDrawer() {
+  platformsDrawer.value = false;
+  settingsDrawer.value = false;
+  actionsDrawer.value = !actionsDrawer.value;
+}
+
 onBeforeUnmount(() => {
   socket.off("scan:scanning_platform");
   socket.off("scan:scanning_rom");
@@ -110,9 +148,140 @@ onMounted(() => {
 <template>
   <v-app>
     <v-main>
+      <v-progress-linear
+        color="romm-accent-1"
+        :active="scanning"
+        :indeterminate="true"
+        absolute
+      />
+
       <notification />
-      <!-- <notification-stack /> -->
+      <!-- TODO: refactor and extract components -->
+      <v-navigation-drawer
+        :floating="platformsDrawer || settingsDrawer"
+        rail
+        rail-width="60"
+      >
+        <template #prepend>
+          <v-row no-gutters class="my-2 justify-center">
+            <router-link :to="{ name: 'dashboard' }">
+              <v-hover v-slot="{ isHovering, props: hoverProps }">
+                <romm-iso
+                  v-bind="hoverProps"
+                  :class="{ 'border-romm-accent-1': isHovering }"
+                  :size="40"
+                />
+              </v-hover>
+            </router-link>
+          </v-row>
+          <v-row no-gutters class="my-4 justify-center">
+            <v-btn
+              v-if="auth.scopes.includes('roms.write')"
+              icon="mdi-upload"
+              variant="flat"
+              @click="emitter?.emit('showUploadRomDialog', null)"
+            />
+            <v-btn
+              icon="mdi-magnify"
+              variant="flat"
+              @click="emitter?.emit('showSearchRomDialog', null)"
+            />
+          </v-row>
+        </template>
+        <v-row no-gutters class="justify-center">
+          <v-btn
+            block
+            rounded="0"
+            variant="flat"
+            :color="platformsDrawer ? 'terciary' : ''"
+            icon
+            @click="togglePlatformsDrawer"
+            ><v-icon :color="$route.name == 'platform' ? 'romm-accent-1' : ''"
+              >mdi-controller</v-icon
+            ></v-btn
+          >
+          <v-btn
+            v-if="auth.scopes.includes('platforms.write')"
+            block
+            rounded="0"
+            variant="flat"
+            color="primary"
+            icon
+            @click="goScan"
+            ><v-icon :color="$route.name == 'scan' ? 'romm-accent-1' : ''"
+              >mdi-magnify-scan</v-icon
+            ></v-btn
+          >
+        </v-row>
+        <template #append>
+          <v-row no-gutters class="justify-center">
+            <v-btn
+              block
+              rounded="0"
+              variant="flat"
+              :color="settingsDrawer ? 'terciary' : ''"
+              icon
+              @click="toggleSettingsDrawer"
+              ><v-icon
+                :color="
+                  $route.name == 'settings' ||
+                  $route.name == 'management' ||
+                  $route.name == 'administration'
+                    ? 'romm-accent-1'
+                    : ''
+                "
+                >mdi-cog</v-icon
+              ></v-btn
+            >
+            <v-avatar size="35" class="my-2">
+              <v-img
+                :src="
+                  auth.user?.avatar_path
+                    ? `/assets/romm/assets/${auth.user?.avatar_path}`
+                    : defaultAvatarPath
+                "
+              />
+            </v-avatar>
+          </v-row>
+        </template>
+      </v-navigation-drawer>
+
       <router-view />
+
+      <v-bottom-navigation
+        v-if="mdAndDown"
+        :elevation="0"
+        class="bg-primary"
+        mode="shift"
+        height="45"
+      >
+        <v-btn @click="goHome">
+          <v-icon>mdi-home</v-icon>
+          <span>Home</span>
+        </v-btn>
+        <v-btn @click="togglePlatformsDrawer">
+          <v-icon>mdi-controller</v-icon>
+          <span>Platforms</span>
+        </v-btn>
+        <v-btn v-if="auth.scopes.includes('platforms.write')" @click="goScan">
+          <v-icon>mdi-magnify-scan</v-icon>
+          <span>Scan</span>
+        </v-btn>
+        <v-btn @click="toggleSettingsDrawer">
+          <v-icon>mdi-cog</v-icon>
+          <span>Settings</span>
+        </v-btn>
+        <v-btn @click="toggleActionsDrawer">
+          <v-icon>mdi-dots-horizontal</v-icon>
+          <span>Actions</span>
+        </v-btn>
+      </v-bottom-navigation>
     </v-main>
   </v-app>
 </template>
+<style scoped>
+.v-progress-linear,
+.v-bottom-navigation {
+  z-index: 9999 !important;
+}
+</style>
